internal/cli/internal/apps: split to multiple files per sub-command
