+++ internal/cli/internal/apps/destroy.go
@@ -1,0 +1,26 @@
+package apps
+
+import (
+	"context"
+
+	"github.com/spf13/cobra"
+
+	"github.com/superfly/flyctl/internal/cli/internal/command"
+	"github.com/superfly/flyctl/internal/cli/internal/flag"
+)
+
+func newDestroy() *cobra.Command {
+	destroy := command.FromDocstrings("apps.destroy", runDestroy,
+		command.RequireSession)
+
+	destroy.Args = cobra.ExactArgs(1)
+
+	flag.Add(destroy, nil,
+		flag.Yes())
+
+	return destroy
+}
+
+func runDestroy(ctx context.Context) error {
+	return command.ErrNotImplementedYet
+}
